+++ services/objectstorage/CHANGELOG.md
@@ -1,2 +1,5 @@
+## v0.1.1
+- Bump dependency `cloud.stackit.sdk.core` to v0.4.1
+
 ## v0.1.0
 - Initial onboarding of STACKIT Java SDK for Object storage service